Keep `CrashReportingFeature` internal
